\name{anchap-package}
\alias{anchap-package}
\alias{anchap}
\docType{package}
\title{
Phasing - uncovering haplotypes from genotypes - in isolated populations.
}
\description{
First establishes pair-wise IBD sharing between individuals in the cohort. 
Then for each individual performs stitching and phasing. 
Stitching is alignment of the shared sequences for an individual into two categories, 
so that in each there are sequences sharing with one haplotype. 
Phasing uses the shared sequences, and their genotype, to recover haplotypes.
Finally, the procedure is evaluated using the parent-offspring trios present in the cohort.
}
\details{
\tabular{ll}{
Package: \tab anchap\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-08-16\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Dominik Glodzik

Maintainer: Who to complain to <yourfault@somewhere.net>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
