\name{computeSurrDensity}
\alias{calcCoverage}
\title{
Compute IBD density
}
\description{
Computes sharing density from sharing encoded as a data frame.
}
\usage{
computeSurrDensity(surrMatDf, noSnps=NA)
}
\arguments{
  \item{surrMatDf}{data frame storing the shared sequences}
  \item{noSnps}{the}
}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}

\seealso{
}
\examples{
}
\keyword{ density }
\keyword{ surrDensity }
